review comments II
* move merge related code out of sync & data mods into new merge mod
