+++ xayn-ai/src/coi/mod.rs
@@ -1,8 +1,10 @@
 mod config;
+mod merge;
 mod system;
 mod utils;
 
 pub(crate) use config::Configuration;
+pub(crate) use merge::reduce_cois;
 pub(crate) use system::CoiSystem;
 
 #[cfg(test)]
