+++ xayn-ai/src/reranker/sync.rs
@@ -1,15 +1,12 @@
 use crate::{
+    coi::reduce_cois,
     data::{CoiPoint, UserInterests},
-    embedding::utils::l2_distance,
     error::Error,
-    utils::nan_safe_f32_cmp_high,
-    CoiId,
 };
 use anyhow::bail;
-use itertools::Itertools;
+use itertools::izip;
 use serde::{Deserialize, Serialize};
 
-const MERGE_THRESHOLD_DIST: f32 = 4.5;
 const CURRENT_SCHEMA_VERSION: u8 = 0;
 
 /// Synchronizable data of the reranker.
@@ -59,101 +56,36 @@ impl SyncData {
     }
 }
 
-/// A pair of CoIs.
-#[derive(Clone)]
-struct CoiPair<C>(C, C);
+impl UserInterests {
+    /// Moves all user interests of `other` into `Self`.
+    pub(crate) fn append(&mut self, mut other: Self) {
+        // TODO drop dupes in other
+        append_cois(&mut self.positive, &mut other.positive);
+        append_cois(&mut self.negative, &mut other.negative);
+    }
 
-impl<C> CoiPair<C>
+    /// Re-assigns CoI ids for normalization.
+    pub(crate) fn reassign_ids(&mut self) {
+        reassign_coi_ids(&mut self.positive);
+        reassign_coi_ids(&mut self.negative);
+    }
+}
+
+fn append_cois<C>(locals: &mut Vec<C>, remotes: &mut Vec<C>)
 where
     C: CoiPoint,
 {
-    /// Merges the CoI pair, assigning it the smaller of the two ids.
-    fn merge(self) -> C {
-        let CoiId(id1) = self.0.id();
-        let CoiId(id2) = self.1.id();
-        self.0.merge(self.1, id1.min(id2))
-    }
+    // shift remote ids to avoid clashes with local ids
+    let max_local_id = locals.iter().map(|coi| coi.id().0).max().unwrap_or(0);
+    remotes
+        .iter_mut()
+        .for_each(|coi| coi.set_id(max_local_id + coi.id().0));
 
-    /// True iff either CoI has the given id.
-    fn contains(&self, id: CoiId) -> bool {
-        self.0.id() == id || self.1.id() == id
-    }
-
-    /// True iff either CoI is one of the given pair.
-    fn contains_any(&self, other: &Self) -> bool {
-        self.contains(other.0.id()) || self.contains(other.1.id())
-    }
+    locals.append(remotes);
 }
 
-/// A `Coiple` is a pair of CoIs and the distance between them.
-#[derive(Clone)]
-struct Coiple<C> {
-    cois: CoiPair<C>,
-    dist: f32,
-}
-
-impl<C> Coiple<C> {
-    fn new(coi1: C, coi2: C, dist: f32) -> Self {
-        let cois = CoiPair(coi1, coi2);
-        Self { cois, dist }
-    }
-}
-
-/// Computes the l2 distance between two CoI points.
-fn dist<C>(coi1: &C, coi2: &C) -> f32
-where
-    C: CoiPoint,
-{
-    l2_distance(coi1.point(), coi2.point())
-}
-
-/// Reduces the given collection of CoIs by successively merging the pair in closest
-/// proximity to each other.
-///
-/// <https://xainag.atlassian.net/wiki/spaces/XAY/pages/2029944833/CoI+synchronisation>
-/// outlines the core of the algorithm.
-fn reduce_cois<C>(cois: &mut Vec<C>)
-where
-    C: CoiPoint + Clone,
-{
-    // initialize collection of close coiples
-    let cois_iter = cois.iter();
-    let mut coiples = cois_iter
-        .clone()
-        .cartesian_product(cois_iter)
-        .filter(|(coi1, coi2)| coi1.id() < coi2.id())
-        .filter_map(|(coi1, coi2)| {
-            let dist = dist(coi1, coi2);
-            (dist < MERGE_THRESHOLD_DIST).then(|| Coiple::new(coi1.clone(), coi2.clone(), dist))
-        })
-        .collect_vec();
-
-    while !coiples.is_empty() {
-        // find the minimum i.e. closest pair of cois
-        let min_pair = coiples
-            .iter()
-            .cloned()
-            .min_by(|cpl1, cpl2| nan_safe_f32_cmp_high(&cpl1.dist, &cpl2.dist))
-            .unwrap() // safe: nonempty coiples
-            .cois;
-
-        // discard pair from collections
-        cois.retain(|coi| !min_pair.contains(coi.id()));
-        coiples.retain(|cpl| !cpl.cois.contains_any(&min_pair));
-
-        let merged_coi = min_pair.merge();
-
-        // record close coiples featuring the merged coi
-        let mut new_coiples = cois
-            .iter()
-            .filter_map(|coi| {
-                let dist = dist(&merged_coi, coi);
-                (dist < MERGE_THRESHOLD_DIST)
-                    .then(|| Coiple::new(merged_coi.clone(), coi.clone(), dist))
-            })
-            .collect_vec();
-        coiples.append(&mut new_coiples);
-
-        cois.push(merged_coi);
+fn reassign_coi_ids(cois: &mut Vec<impl CoiPoint>) {
+    for (id, coi) in izip!(1..cois.len() + 1, cois) {
+        coi.set_id(id)
     }
 }
